requirements.txt
absl-py==0.12.0
astor==0.8.1
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
bleach==3.3.0
certifi==2021.5.30
cffi==1.14.5
chardet==3.0.4
click==7.1.2
cmake==3.22.0
colorlover==0.3.0
cycler==0.10.0
cymem==2.0.5
Cython==0.29.23
decorator==4.4.2
defusedxml==0.7.1
Flask==1.1.4
gym==0.17.3
h5py==3.1.0
hdfs==2.6.0
imageio==2.4.1
imagesize==1.2.0
imbalanced-learn==0.4.3
imblearn==0.0
ipykernel==4.10.1
ipython==5.5.0
ipython-genutils==0.2.0
jedi==0.18.0
jieba==0.42.1
Jinja2==2.11.3
joblib==1.0.1
Keras==2.4.3
Keras-Preprocessing==1.1.2
lvis==0.5.3
lxml==4.2.6
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.2.2
numpy==1.19.5
packaging==20.9
palettable==3.3.0
pandas==1.1.5
param==1.10.1
parso==0.8.2
pathlib==1.0.1
patsy==0.5.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
pip-tools==4.5.1
prometheus-client==0.11.0
prompt-toolkit==1.0.18
protobuf==3.12.4
pycparser==2.20
pyct==0.4.8
Pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2018.9
pyviz-comms==2.0.2
pyzmq==22.1.0
requests==2.23.0
Send2Trash==1.5.0
six==1.15.0
sklearn==0.0
sklearn-pandas==1.8.0
tensorboard==2.5.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
terminado==0.10.1
testpath==0.5.0
tornado==5.1.1
tqdm==4.41.1
urllib3==1.24.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1